gcp-helpers/vertex-ai-feature-store/tools/requirements.txt
# Vertex AI Feature Store Dependencies
# Core Google Cloud AI Platform SDK and Vertex AI SDK
google-cloud-aiplatform>=1.38.0
google-cloud-aiplatform[vertex-ai]>=1.38.0

# Authentication and Core Google Cloud libraries
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0

# Additional Google Cloud dependencies
google-api-core>=2.11.0
google-cloud-core>=2.3.0

# Protocol Buffers (required for API communication)
protobuf>=4.21.0

# HTTP and networking
requests>=2.28.0
urllib3>=1.26.0

# Date/time handling
python-dateutil>=2.8.0

# Logging and utilities (optional but recommended)
structlog>=22.3.0
